Resolve #19 add-macos-text-insertion: paste with restore on macOS (#34)
* Plan add-macos-text-insertion (#19)

The OpenSpec change for paste with restore on macOS: proposal, design,
the text-insertion spec delta and tasks.


* Add paste with restore on macOS

The macOS implementations behind the shared TextInserter: the pasteboard
through the Swift helper on its own thread (read only while macOS allows
it without asking, nspasteboard.org markers, current host only, file
promises left out), CGEvent keystrokes with the layout's key of V and
typed text in 20-unit chunks, the focused window through the
Accessibility API, and the new SecureInputOn outcome at the final gate.
The helper's ABI goes to 3. AddTextInsertion() now runs on macOS too;
nothing there calls the inserter until add-macos-dictation.


* Tick task 5.2: CI passed on Windows and macOS


---------

+++ src/Pisum.Transcribe/TextInsertion/Windows/Win32ClipboardService.cs
@@ -91,7 +91,7 @@ internal sealed class Win32ClipboardService : IClipboardService, IDisposable
     }
 
     /// <inheritdoc />
-    public uint SequenceNumber => PInvoke.GetClipboardSequenceNumber();
+    public long SequenceNumber => PInvoke.GetClipboardSequenceNumber();
 
     /// <inheritdoc />
     public Task<ClipboardSnapshot?> TrySnapshotAsync()
